backend/requirements.txt
fastapi
uvicorn
requests
httpx
python-dotenv

SQLAlchemy
psycopg2-binary
pgvector

pydantic
pydantic-settings

# torch==2.2.2+cpu --extra-index-url https://download.pytorch.org/whl/cpu
sentence-transformers

groq

dateparser
pytz

numpy
scipy
scikit-learn
 
chromadb
loguru
gspread
oauth2client
pandas
openai
